+++ src/components/Prologue.vue
@@ -1,0 +1,13 @@
+<script setup lang="ts">
+import SectionTitle from './SectionTitle.vue'
+</script>
+
+<template>
+  <aside class="mb-5 p-5 border-4 border-amber-400 text-black dark:text-white">
+    <SectionTitle anchor="prologue" title="⚠️ Experimental" />
+    <div class="space-y-2">
+      <p><strong>This tool is an experiment and should not be relied upon.</strong></p>
+      <p>If it proves useful then it may form the basis of a more official set of tools.</p>
+    </div>
+  </aside>
+</template>